requirements.txt
matplotlib
pandas
scikit-learn
numpy
python-levenshtein
nltk
random